Delete mapper test stuff and move mapper package
